+++ projects/ng-dynamic-json-form/src/lib/components/form-title/form-title.component.ts
@@ -33,6 +33,7 @@ export class FormTitleComponent {
 
   @Input() label?: string;
   @Input() layout?: FormControlConfig['layout'];
+  @Input() extra?: FormControlConfig['extra'];
   @Input() collapsibleEl?: HTMLElement;
   @Input() state?: FormLayout['contentCollapsible'];
   @Input() customComponent?: LayoutComponents['formTitle'];
@@ -101,6 +102,7 @@ export class FormTitleComponent {
     );
     componentRef.instance.label = this.label;
     componentRef.instance.layout = this.layout;
+    componentRef.instance.extra = this.extra;
     componentRef.instance.collapsibleEl = this.collapsibleEl;
   }